killer moves (-40% nodes reduction!!)

+++ src/board.rs
@@ -441,7 +441,7 @@ impl Board {
         }
 
         board.pos.rule_fifty = half_move_str.parse::<u8>().unwrap();
-        board.pos.ply = full_move_str.parse::<u16>().unwrap();
+        board.pos.ply = full_move_str.parse::<usize>().unwrap();
 
         let mut file = 0;
         let mut rank = 7;
